chore: fix unit and integration tests
